moved libp2p code to sdk examples
