--- clients/libp2p/Dockerfile.build-nym
@@ -1,22 +1,0 @@
-FROM rust:1.67 as builder
-RUN apt-get install -y pkg-config git 
-
-WORKDIR /nym
-# Build the nym binaries locally.
-RUN git clone https://github.com/nymtech/nym && cd nym && git checkout release/v1.1.12
-RUN cd nym && cargo build --release
-RUN mv ./nym/target/release/nym-client ./nym-client
-
-FROM ubuntu:focal as main
-RUN apt-get update -qqq
-RUN apt-get install -y wget ca-certificates build-essential
-RUN apt-get install libssl1.1
-WORKDIR /nym
-COPY --from=builder /nym/nym-client ./
-RUN chmod a+rx ./nym-client && mv ./nym-client /usr/bin/
-ENV NYM_ID 'default'
-COPY entrypoint.sh /usr/bin/entrypoint.sh
-RUN chmod +x /usr/bin/entrypoint.sh
-EXPOSE 1977
-# ENTRYPOINT ["entrypoint.sh"]
-CMD ["entrypoint.sh"]
