--- clients/libp2p/Dockerfile.nym
@@ -1,14 +1,0 @@
-FROM ubuntu:focal
-RUN apt-get update -qqq
-RUN apt-get install -y wget ca-certificates build-essential
-RUN apt-get install libssl1.1
-WORKDIR /nym
-# Get the nym binaries from github.
-RUN wget https://github.com/nymtech/nym/releases/download/nym-binaries-v1.1.12/nym-client
-RUN chmod a+rx nym-client && mv nym-client /usr/bin/
-ENV NYM_ID 'default'
-COPY entrypoint.sh /usr/bin/entrypoint.sh
-RUN chmod +x /usr/bin/entrypoint.sh
-EXPOSE 1977
-# ENTRYPOINT ["entrypoint.sh"]
-CMD ["entrypoint.sh"]
